update to deleteOrg
deleteOrganization should now remove the itself from the accounts that have it interested in it

+++ data/organizations.js
@@ -428,6 +428,15 @@ const organizationFunctions ={
         o_id= await id_validation.checkOrganizationID(o_id);
         const organizationCollection= await organizations();
         if(!organizationCollection) throw 'Failed to connect to organization collection'; 
+        //remove account's interest in organization through iterating over interestedAccounts list and calling removeInterestedOrg in accountsFunctions
+        //find the organization to be deleted
+        const organization = await organizationCollection.findOne({_id: new ObjectId(o_id)});
+        if(!organization) throw 'Organization does not exist';
+        //iterate over interestedAccounts list and call removeInterestedOrg
+        organization.interestedAccounts.forEach(async (account_id) => {
+            await accountsFunctions.removeInterestedOrg(account_id, o_id);
+        });
+        //finally delete the organization
         const organizationData =  await organizationCollection.findOneAndDelete({_id: new ObjectId(o_id)});
         if(!organizationData) throw 'Cannot delete organization';
         return `${organizationData.name} have been successfully deleted!`;
